Refactor category management in newExpenseForm and myBudgets components
- Updated `newExpenseForm.vue` to utilize `categories` from `useCategoryStore` instead of `getCategories` from `useMyExpensesStore`, enhancing clarity and consistency in category selection.
- Modified `myBudgets.vue` to fetch categories from `useCategoryStore`, streamlining category management alongside budget operations.
- Removed redundant category fetching logic from `expensesStore`, improving state management and modularity.
